Refactoring to break down exports and visulization into new files
