+++ modules/tests/test_crawler.py
@@ -142,96 +142,6 @@ class TestCrawlerFunctions(unittest.TestCase):
         self.assertIn("https://torcrawl.com/deep/custom-path", result)
         self.assertIn("https://torcrawl.com/deep/second-path", result)
 
-    def test_export_findings_creates_json(self):
-        """Findings are exported to JSON with separated sections."""
-        prefix = f"{self.crawler.timestamp}_results_test"
-        self.crawler.findings["links"].update({"https://torcrawl.com", "https://torcrawl.com/about"})
-        self.crawler.findings["external_links"].add("https://external.com/path")
-        self.crawler.findings["images"].add("https://torcrawl.com/img/logo.png")
-        self.crawler.findings["emails"].add("mailto:test@torcrawl.com")
-
-        self.crawler.export_findings(self.out_path, prefix, export_json=True, export_xml=False)
-
-        json_path = os.path.join(self.out_path, f"{prefix}.json")
-        self.assertTrue(os.path.exists(json_path))
-
-        with open(json_path, "r", encoding="UTF-8") as json_file:
-            payload = json.load(json_file)
-
-        self.assertIn("links", payload)
-        self.assertIn("external_links", payload)
-        self.assertIn("images", payload)
-        self.assertIn("emails", payload)
-        self.assertIn("https://torcrawl.com/about", payload["links"])
-        self.assertIn("https://external.com/path", payload["external_links"])
-
-    def test_export_findings_creates_xml(self):
-        """Findings are exported to XML with separated sections."""
-        prefix = f"{self.crawler.timestamp}_results_test_xml"
-        self.crawler.findings["links"].update({"https://torcrawl.com"})
-        self.crawler.findings["scripts"].add("https://torcrawl.com/static/app.js")
-        self.crawler.findings["telephones"].add("tel:012-013-104-5")
-
-        self.crawler.export_findings(self.out_path, prefix, export_json=False, export_xml=True)
-
-        xml_path = os.path.join(self.out_path, f"{prefix}.xml")
-        self.assertTrue(os.path.exists(xml_path))
-
-        tree = ET.parse(xml_path)
-        root = tree.getroot()
-
-        links_section = root.find("links")
-        scripts_section = root.find("scripts")
-        telephones_section = root.find("telephones")
-
-        self.assertIsNotNone(links_section)
-        self.assertIsNotNone(scripts_section)
-        self.assertIsNotNone(telephones_section)
-        self.assertEqual(links_section[0].text, "https://torcrawl.com")
-        self.assertEqual(scripts_section[0].text, "https://torcrawl.com/static/app.js")
-        self.assertEqual(telephones_section[0].text, "tel:012-013-104-5")
-
-    def test_export_database_stores_nodes_edges_and_titles(self):
-        """SQLite export stores nodes, edges, and titles."""
-        prefix = f"{self.crawler.timestamp}_results_test_db"
-        self.crawler.findings["links"].update({"https://torcrawl.com", "https://torcrawl.com/about"})
-        self.crawler.edges.add(("https://torcrawl.com", "https://torcrawl.com/about"))
-        self.crawler.titles["https://torcrawl.com"] = "Home"
-        self.crawler.titles["https://torcrawl.com/about"] = "About"
-
-        self.crawler.export_database(self.out_path, prefix)
-
-        db_path = os.path.join(self.out_path, f"{prefix}.db")
-        self.assertTrue(os.path.exists(db_path))
-
-        conn = sqlite3.connect(db_path)
-        with conn:
-            cur = conn.cursor()
-            cur.execute("SELECT url, title FROM nodes ORDER BY url;")
-            rows = cur.fetchall()
-            self.assertIn(("https://torcrawl.com", "Home"), rows)
-            self.assertIn(("https://torcrawl.com/about", "About"), rows)
-
-            cur.execute("SELECT from_url, to_url FROM edges;")
-            edges = cur.fetchall()
-            self.assertIn(("https://torcrawl.com", "https://torcrawl.com/about"), edges)
-
-    def test_export_visualization_creates_html(self):
-        """Visualization export produces an HTML graph file."""
-        prefix = f"{self.crawler.timestamp}_results_test_vis"
-        self.crawler.findings["links"].update({"https://torcrawl.com", "https://torcrawl.com/about"})
-        self.crawler.edges.add(("https://torcrawl.com", "https://torcrawl.com/about"))
-        self.crawler.titles["https://torcrawl.com"] = "Home"
-        self.crawler.titles["https://torcrawl.com/about"] = "About"
-
-        self.crawler.export_database(self.out_path, prefix)
-        self.crawler.export_visualization(self.out_path, prefix)
-
-        html_path = os.path.join(self.out_path, f"{prefix}_graph.html")
-        self.assertTrue(os.path.exists(html_path))
-        with open(html_path, "r", encoding="utf-8") as html_file:
-            contents = html_file.read()
-        self.assertIn("<html", contents.lower())
 
     def test_make_request_with_random_ua_and_proxy(self):
         crawler = Crawler("http://example.com", 0, 0, self.out_path, False, False, random_ua=True, random_proxy=True)
